requirements.txt
pyyaml==5.4.1
splunk-appinspect==2.30.0
tabulate==0.9.0